sandbox/modules/identity/src/main/java/org/opensearch/identity/authmanager/noop/NoopAuthenticationManager.java
/*
 * SPDX-License-Identifier: Apache-2.0
 *
 * The OpenSearch Contributors require contributions made to
 * this file be licensed under the Apache-2.0 license or a
 * compatible open source license.
 */

package org.opensearch.identity.authmanager.noop;

import org.opensearch.authn.AccessTokenManager;
import org.opensearch.authn.AuthenticationManager;
import org.opensearch.authn.Subject;

/**
 * Implementation of authentication manager that does not enforce authentication
 *
 * This class and related classes in this package will not return nulls or fail permissions checks
 *
 * @opensearch.internal
 */
public class NoopAuthenticationManager implements AuthenticationManager {

    @Override
    public Subject getSubject() {
        return new NoopSubject();
    }

    @Override
    public AccessTokenManager getAccessTokenManager() {
        return null;
    }
}
